Implemented the requested changes; left some comments
